Add more field getters for raw Slam BivariateMaps
SlamTmplField represents a slam::BivariateMap which is templated on a
ProductSet/RelationSet, whose constituent sets are slam::RangeSets.

SlamTmplStrideField is the same as above, but also has a compile-time
stride of one.

These should help us investigate the impact of using the abstract
slam::Set vs a concrete slam::RangeSet type, as well as using runtime
vs. compile-time stride.