Trivial changes to trigger autotests.
These are needed changes but trivial.  This commit is to force
the CI test.  A recent commit had failure that seemed unrelated
to the bug I'm trying to track down.  It cleared up after repeating
the test.  But just in case, I want to run another round of tests
before making real changes.