Move spacing out of ArrayBase and into the children classes.
This change restores the old optimized stateless 1D ArrayBase.
It also allows Array to hardwire the spacing to 1 so it can be
removed by compiler optimization.  Non-unit spacing is meant for
ArrayViews and not useful for Arrays anyway.  The spacing()
method is implemented in the child classes, in the same way
that data() and size() are.