Multimat: fix segfault in examples/calculate.cpp
In calculate.cpp, we have mappings from virtual interface-based
ProductSets and RelationSets to their corresponding concrete instances.
Keeping those mapping variables as global variables seems to create a
conflict at destruction time with Umpire's static state.

To fix this, we only keep pointers to the mappings as global variables,
and create the underlying mapping variables within the scope of the main()
function.