Locally disables death tests in MPI-enabled configs for signed_distance_interface tests
Gtest changed how they define `GTEST_HAS_DEATH_TEST`
from "always defined with a value of 0 or 1"
to "defined when available or undefined."
This made our previous workaround stop working.

The key problem appears to be that death tests are not properly handled
in MPI configurations, leading to a "broken pipe" error.

See:
* https://github.com/google/googletest/issues/3267
* https://github.com/google/googletest/commit/23142843f7c875907a6ed7985174a6e346e82c3d#diff-97f768d897442b94f24002f6bcff7124e3c73d05b3916fb492ee29436ce0701f