Array: fix reallocation for non-trivially relocatable types
Adds a new helper method ArrayOps::realloc_move, which calls
std::uninitialized_move if the memory is located on the CPU. This
ensures that objects are properly constructed on reallocation.

In the case of device-side memory, we can just continue to do a
bit-copy, since we don't support non-trivially relocatable types on the
device.