Update to Shroud 0.13.0
Changed to use a public install of Shroud in
/collab/usr/gapps/shroud/public/$SYS_TYPE.
Edited the host-configs files to use this version.

I added a new spack/packages/py-shroud/package.py file which includes
version 0.13.0 and tested that it builds properly.
It should be upstreamed into spack.

The Fortran and C wrappers have not been regenerated with this version
yet.

#1258