Array: add combined support for device-only/unified in ArrayOps
As with ArrayBase<Device>, we can defer operations on trivally copyable
and trivially default-constructible types to Umpire, while performing
these operations on the host when construction or copying isn't trivial.

However, unified and pinned memory are accessible on the host, so we can
just directly call the host operation on the memory ranges in question
instead of having to create a temporary host staging buffer.