Fixes memory leaks in MFEMSidreDataCollection and associated tests
The MFEMSidreDataCollection has a flag for whether it owns the data
associated with the mesh, e.g. the high order fields. If so, it deallocates
them upon destruction. It's base class has a separate flag for whether it owns
the mfem mesh (`own_data`). MFEMSidreDataCollection's constructor was not setting
the latter, which lead to some memory leaks in cases where users expected
the data collection to deallocate the mesh. This commit fixes the problem
by setting `mfem::DataCollection::own_data` to the same value as `owns_mesh_data`
in the constructor, and fixes tests where this was not the desired behavior.