Adds coordinate system transform to svg2contours script
Converts from SVG's coordinate system, where the y-axis points down
to a cordinate system in which the y-axis points up.

Specifically, when the SVG has a top-level height attribute,
the new y coordinate is computed from the old one as:
```
  y_new = svg_height - y_old
```

Note: This also requires reversing the orientation of each curve.