FlatMap: changes to support inserting into existing map with gaps
Before, the batched insertion algorithm was able to probe until it
reached an empty bucket. In a map constructed with only insertions, this
is a valid approach since the empty buckets are guaranteed to only occur
at the end of the probing sequence.

If a map contains gaps in probing sequences as a result of erasures, we
must continue probing to find potential duplicates. However, we can
insert the pair as if it was the only key in the first empty bucket we
find, then continue probing. If a subsequent bucket with an equivalent
key is found, we can just delete that bucket.